+++ CHANGELOG.md
@@ -5,8 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.1]
+## [0.2.0] - 2023-07-01
+### Added
+- HICP pre-processing
+- HICP parquet cache saving
 
+### Remove
+- Eurostat cache saving
+
+
+## [0.1.2] - 2023-07-01
+### Changed
+- Eurostat functions logging
+
+## [0.1.1] - 2023-07-01
 ### Added
 - Python >3.13 support
 
@@ -14,7 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `scipy` and `networkx` dependencies
 
 ## [0.1.0] - 2023-07-01
-
 ### Added
 - First release
 - Fully working HICP